+++ test/init/file.ts
@@ -1,0 +1,5 @@
+const val: string = 123;
+
+declare function expectsNumber(v: number): string;
+
+expectsNumber('string').someMethod;